+++ framework/validators/ExistValidator.php
@@ -69,9 +69,9 @@ class ExistValidator extends Validator
      */
     public $targetRelation;
     /**
-     * @var string|array|\Closure additional filter to be applied to the DB query used to check the existence of the attribute value.
+     * @var string|array|callable additional filter to be applied to the DB query used to check the existence of the attribute value.
      * This can be a string or an array representing the additional query condition (refer to [[\yii\db\Query::where()]]
-     * on the format of query condition), or an anonymous function with the signature `function ($query)`, where `$query`
+      * on the format of query condition), or an callable with the signature `function ($query)`, where `$query`
      * is the [[\yii\db\Query|Query]] object that you can modify in the function.
      */
     public $filter;
@@ -125,7 +125,7 @@ class ExistValidator extends Validator
         /** @var ActiveQuery $relationQuery */
         $relationQuery = $model->{'get' . ucfirst($this->targetRelation)}();
 
-        if ($this->filter instanceof \Closure) {
+        if ($this->filter instanceof \Closure || (is_array($this->filter) && is_callable($this->filter))) {
             call_user_func($this->filter, $relationQuery);
         } elseif ($this->filter !== null) {
             $relationQuery->andWhere($this->filter);
@@ -295,7 +295,7 @@ class ExistValidator extends Validator
     {
         /* @var $targetClass \yii\db\ActiveRecordInterface */
         $query = $targetClass::find()->andWhere($condition);
-        if ($this->filter instanceof \Closure) {
+        if ($this->filter instanceof \Closure || (is_array($this->filter) && is_callable($this->filter))) {
             call_user_func($this->filter, $query);
         } elseif ($this->filter !== null) {
             $query->andWhere($this->filter);